+++ CHANGELOG.rst
@@ -1,3 +1,9 @@
+2.0.0 (Jun XX, 2016)
+--------------------
+New feature: *Edit related* buttons next to *add another* buttons, to allow for quick
+modification of related objects in a popup. **Backwards incompatible:** ``PopupMixin``
+has been renamed to ``AddPopupMixin`` without a deprecation path.
+
 1.0.0 (Mar 17, 2016)
 --------------------
 Rewrite/merge with @yourlabs implementation. Not backwards compatible!